사용자제어_작업진행중

+++ backend/src/main/java/com/jibangyoung/domain/admin/dto/AdRegionDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AdRegionDTO {
-    private Long region_code;
+    private Integer region_code;
     private String sido; 
 }
